fix: move indicators.py to init
